Accessibility updates
Changed logout button to not respond to enter key presses in other fields, such as the employee search page.
Changed CSS to be WCAG 2.0 AAA compliant.
Fixed missing end tags on a few pages.
Fixed incorrectly spelled tags, such as <aabr> instead of <abbr>, or a </td> closing a <th>.
Moved spans with onclick events for confirmations to the OnClientClick in the button that they are wrapped around. This allows the button to be navigated to by the keyboard.
Added <thead> to table headers.